Move setup utilities into common module to share between test modules
