+++ crates/hotfix/tests/common/mod.rs
@@ -1,4 +1,5 @@
 pub mod mock_application;
 pub mod mock_counterparty;
 pub mod session_assertions;
+pub mod setup;
 pub mod test_messages;
